Backend/requirements.txt
anyio==4.8.0
base58==2.1.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
ECPy==1.2.5
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
pycryptodome==3.22.0
sniffio==1.3.1
types-Deprecated==1.2.15.20250304
typing_extensions==4.12.2
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xrpl-py==4.1.0
